Actually commenting on iNat observations with a summary of data.

+++ server/tools/inaturalist/main.go
@@ -1,17 +1,25 @@
 package main
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"flag"
+	"fmt"
+	"math/rand"
+	"net/http"
 	"time"
 
 	"go.uber.org/zap"
 
 	"github.com/Conservify/gonaturalist"
 
+	fk "github.com/fieldkit/cloud/server/api/client"
+	fktesting "github.com/fieldkit/cloud/server/tools"
+
+	"github.com/fieldkit/cloud/server/backend/ingestion/formatting"
 	"github.com/fieldkit/cloud/server/inaturalist"
 	"github.com/fieldkit/cloud/server/logging"
-	fktesting "github.com/fieldkit/cloud/server/tools"
 )
 
 var (
@@ -28,9 +36,12 @@ type options struct {
 	Expedition string
 	DeviceName string
 	DeviceId   string
+
+	DeleteComments   bool
+	NumberOfReadings int
 }
 
-func findObservation(ctx context.Context, log *zap.SugaredLogger, inc *gonaturalist.Client) (obs *gonaturalist.SimpleObservation, err error) {
+func findObservation(ctx context.Context, log *zap.SugaredLogger, o *options, inc *gonaturalist.Client) (obs *gonaturalist.SimpleObservation, err error) {
 	user, err := inc.GetCurrentUser()
 	if err != nil {
 		return nil, err
@@ -68,7 +79,7 @@ func findObservation(ctx context.Context, log *zap.SugaredLogger, inc *gonatural
 	return myObservations.Observations[0], nil
 }
 
-func deleteMyOwnComments(ctx context.Context, log *zap.SugaredLogger, inc *gonaturalist.Client, obs *gonaturalist.SimpleObservation) (err error) {
+func deleteMyOwnComments(ctx context.Context, log *zap.SugaredLogger, o *options, inc *gonaturalist.Client, obs *gonaturalist.SimpleObservation) (err error) {
 	user, err := inc.GetCurrentUser()
 	if err != nil {
 		return err
@@ -82,10 +93,12 @@ func deleteMyOwnComments(ctx context.Context, log *zap.SugaredLogger, inc *gonat
 
 	for _, c := range full {
 		if c.UserId == user.Id {
-			log.Infow("Deleting", "comment", c)
-			err = inc.DeleteComment(c.Id)
-			if err != nil {
-				return err
+			if o.DeleteComments {
+				log.Infow("Deleting", "comment", c)
+				err = inc.DeleteComment(c.Id)
+				if err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -93,14 +106,75 @@ func deleteMyOwnComments(ctx context.Context, log *zap.SugaredLogger, inc *gonat
 	return nil
 }
 
-func updateObservation(ctx context.Context, log *zap.SugaredLogger, inc *gonaturalist.Client, obs *gonaturalist.SimpleObservation) (err error) {
+func updateObservation(ctx context.Context, log *zap.SugaredLogger, o *options, inc *gonaturalist.Client, obs *gonaturalist.SimpleObservation) (err error) {
+	now := time.Now()
 	updateObservation := gonaturalist.UpdateObservationOpt{
-		Id:          obs.Id,
-		Description: obs.Description + "\n" + time.Now().String(),
+		Id:               obs.Id,
+		ObservedOnString: &now,
+		Description:      obs.Description + "\n" + time.Now().String(),
 	}
 	err = inc.UpdateObservation(&updateObservation)
 
 	return err
+}
+
+func addFakeRecords(ctx context.Context, log *zap.SugaredLogger, o *options, fkc *fk.Client) error {
+	for i := 0; i < o.NumberOfReadings; i += 1 {
+		m := formatting.HttpJsonMessage{
+			Location: GreekTheaterLocation,
+			Time:     time.Now().UTC().Unix(),
+			Device:   o.DeviceId,
+			Fixed:    true,
+			Stream:   "",
+			Modules:  []string{"Fake"},
+			Values: map[string]interface{}{
+				"cpu":      rand.Uint32() % 100,
+				"temp":     rand.Uint32() % 40,
+				"humidity": rand.Uint32() % 100,
+			},
+		}
+
+		log.Infow("Adding fake reading", "deviceId", o.DeviceId)
+
+		bytes, err := json.Marshal(m)
+		if err != nil {
+			log.Fatalf("Error %v", err)
+		}
+
+		err = o.postRawMessage(bytes)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func triggerRefresh(ctx context.Context, log *zap.SugaredLogger, o *options) error {
+	url := o.getUrl("/tasks/five")
+	log.Infow("Refreshing", "url", url)
+	_, err := http.Get(url)
+	return err
+}
+
+func (o *options) getUrl(path string) string {
+	return fmt.Sprintf("%s://%s%s", o.Scheme, o.Host, path)
+}
+
+func (o *options) postRawMessage(body []byte) error {
+	url := o.getUrl("/messages/ingestion")
+	if false {
+		fmt.Printf("curl -X POST -H 'Content-Type: %s' -d '%s' %s\n", formatting.HttpProviderJsonContentType, string(body), url)
+	} else {
+		bytes := bytes.NewBufferString(string(body))
+		url := url
+		url += "?token=" + "IGNORED"
+		_, err := http.Post(url, formatting.HttpProviderJsonContentType, bytes)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 func main() {
@@ -115,6 +189,9 @@ func main() {
 	flag.StringVar(&o.Expedition, "expedition", "", "expedition")
 	flag.StringVar(&o.DeviceName, "device-name", "fkn-test", "device name")
 	flag.StringVar(&o.DeviceId, "device-id", "1c7c6a39-4208-4978-811f-5cc75c83ca13", "device name")
+
+	flag.BoolVar(&o.DeleteComments, "delete-comments", false, "delete comments")
+	flag.IntVar(&o.NumberOfReadings, "number-of-readings", 10, "number of readings")
 
 	flag.Parse()
 
@@ -136,35 +213,33 @@ func main() {
 	}
 	log.Infow("Device", "device", device)
 
-	config := inaturalist.ConfigLocalhost
+	config := inaturalist.ConfigConservify
 	authenticator := gonaturalist.NewAuthenticatorAtCustomRoot(config.ApplicationId, config.Secret, config.RedirectUrl, config.RootUrl)
 	inc := authenticator.NewClientWithAccessToken(config.AccessToken, &gonaturalist.NoopCallbacks{})
 
-	obs, err := findObservation(ctx, log, inc)
+	obs, err := findObservation(ctx, log, &o, inc)
 	if err != nil {
 		panic(err)
 	}
 
-	err = deleteMyOwnComments(ctx, log, inc, obs)
+	err = deleteMyOwnComments(ctx, log, &o, inc, obs)
 	if err != nil {
 		panic(err)
 	}
 
-	err = updateObservation(ctx, log, inc, obs)
+	err = updateObservation(ctx, log, &o, inc, obs)
 	if err != nil {
 		panic(err)
 	}
 
-	if false {
-		addComment := gonaturalist.AddCommentOpt{
-			ParentType: gonaturalist.Observation,
-			ParentId:   obs.Id,
-			Body:       "Hello, world!",
-		}
-		err = inc.AddComment(&addComment)
-		if err != nil {
-			panic(err)
-		}
+	err = addFakeRecords(ctx, log, &o, fkc)
+	if err != nil {
+		panic(err)
+	}
+
+	err = triggerRefresh(ctx, log, &o)
+	if err != nil {
+		panic(err)
 	}
 
 	_ = fkc
